Extract core assertion handling to doctest/parts/assert/handler.{h,cpp}
