+++ CMakeLists.txt
@@ -68,6 +68,7 @@ if(${DOCTEST_WITH_MAIN_IN_STATIC_LIB})
         ${doctest_parts_folder}/private/doctest.cpp
         ${doctest_parts_folder}/private/assert/data.cpp
         ${doctest_parts_folder}/private/assert/expression.cpp
+        ${doctest_parts_folder}/private/assert/handler.cpp
         ${doctest_parts_folder}/private/assert/result.cpp
         ${doctest_parts_folder}/private/assert/type.cpp
         ${doctest_parts_folder}/private/debugger.cpp
